RFCT Compute features in a single Python script
This avoids manipulating/re-manipulating the TSV files

The end result is also significantly faster (while additionally avoiding
unnecessary I/O): down to 3.5s (from ~6s before), although this is not a
significant time sink in the overall FACS pipeline.
